requirements.txt
midv500
pillow
torch 
torchvision
